requirements.txt
discord.py==2.5.0
aiohttp==3.11.12
pytz==2023.3.post1
python-dotenv==1.0.0
aiohappyeyeballs==2.4.6
aiosignal==1.3.2
asyncio==3.4.3
attrs==25.1.0
frozenlist==1.5.0
idna==3.10 
# dev-test
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0 